# Save 17 seconds: Optimize Jinja2 template rendering (21% faster) for relay-list